commit some file

+++ src/main/java/com/poly/spring5mvc/controller/DepartmentController.java
@@ -14,7 +14,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.poly.spring5mvc.model.Departments;
+import com.poly.spring5mvc.model.Staffs;
 import com.poly.spring5mvc.service.DepartmentService;
+import com.poly.spring5mvc.service.StaffService;
 
 @Controller
 public class DepartmentController {
@@ -25,8 +27,13 @@ public class DepartmentController {
 	@Autowired
 	DepartmentService DepartmentService;
 
+	@Autowired
+	StaffService StaffService;
+
 	Boolean saveSucessfull = false;
 	Boolean deleteSucessfull = false;
+	Boolean notDelete = false;
+	
 
 	// Lấy toàn phòng ban (done)
 	@RequestMapping("/department")
@@ -34,9 +41,11 @@ public class DepartmentController {
 		if (saveSucessfull) {
 			model.addAttribute("message", "Thêm thành công!");
 		}
-
 		if (deleteSucessfull) {
 			model.addAttribute("message", "Xóa thành công!");
+		}
+		if (notDelete) {
+			model.addAttribute("message", "Có nhân viên trong phòng ban. Không thể xóa!");
 		}
 		model.addAttribute("department", new Departments());
 
@@ -46,6 +55,7 @@ public class DepartmentController {
 		model.addAttribute("departList", depart);
 		saveSucessfull = false;
 		deleteSucessfull = false;
+		notDelete = false;
 		return "departmentView/department";
 	}
 
@@ -61,7 +71,7 @@ public class DepartmentController {
 	// nhận request và lưu phòng ban (done)
 	@RequestMapping(value = "save-department")
 	public String saveDepartment(ModelMap model, @RequestParam("nameDepart") String nameDepartment) {
-		
+
 		Departments depart = new Departments();
 		depart.setDepartsName(nameDepartment);
 		depart.setRewardPoint(0);
@@ -73,8 +83,15 @@ public class DepartmentController {
 
 	// xóa phòng ban(done)
 	@RequestMapping(value = "/delete-department", method = RequestMethod.GET)
-	public String deteteStaff(@RequestParam("departsId") int departsId) {
+	public String deteteStaff(ModelMap model, @RequestParam("departsId") int departsId) {
+		List<Staffs> staffs = 	StaffService.checkIsExistsStaffOfDepartment(departsId);
 		
+		if (staffs.isEmpty() == false) {
+			System.out.println("lisst dduwojc goi trong if" + StaffService.checkIsExistsStaffOfDepartment(departsId).get(0).getEmail());
+			System.out.println("Không thể xóa vì đã có nhân viên trong phòng ban này");
+			notDelete = true;
+			return ("redirect: department");
+		}
 		DepartmentService.deleteDepartment(departsId);
 		deleteSucessfull = true;
 		return ("redirect: department");
